tests/requirements.txt
# Test dependencies for SD Elements MCP Server
# These are NOT required to run the MCP server itself
# Only needed for running tests

# Core testing framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Integration test dependencies (OpenAI client)
# Only needed for integration tests that verify prompt-to-tool mapping
openai>=1.0.0

# Environment variable loading
python-dotenv>=1.0.0

